refactor: use cls for status history mixin
